+++ test/ash_metrics/backend_test.exs
@@ -16,19 +16,6 @@ defmodule AshMetrics.BackendTest.Reporter do
   end
 end
 
-defmodule AshMetrics.BackendTest.GaugeReporter do
-  @moduledoc false
-  # A backend that reports the gauges itself.
-
-  @behaviour AshMetrics.Backend
-
-  @impl AshMetrics.Backend
-  def child_spec(_opts), do: :ignore
-
-  @impl AshMetrics.Backend
-  def polls_gauges?, do: true
-end
-
 defmodule AshMetrics.BackendTest.ValueReporter do
   @moduledoc false
   # A backend that takes the gauge values a poll collects.
@@ -47,7 +34,6 @@ defmodule AshMetrics.BackendTest do
   use ExUnit.Case, async: false
 
   alias AshMetrics.Backend
-  alias AshMetrics.BackendTest.GaugeReporter
   alias AshMetrics.BackendTest.Reporter
   alias AshMetrics.BackendTest.ValueReporter
   alias AshMetrics.Test.Invoice
@@ -101,24 +87,6 @@ defmodule AshMetrics.BackendTest do
       assert is_pid(pid)
 
       Supervisor.stop(supervisor)
-    end
-  end
-
-  describe "polls_gauges?/0" do
-    test "is false for a backend that does not implement the callback" do
-      Application.put_env(:ash_metrics, :backend, Reporter)
-
-      refute Backend.polls_gauges?()
-    end
-
-    test "is false for the default backend" do
-      refute Backend.polls_gauges?()
-    end
-
-    test "is what the backend says" do
-      Application.put_env(:ash_metrics, :backend, GaugeReporter)
-
-      assert Backend.polls_gauges?()
     end
   end